Moved models to monitor/api
Also fixed a few review comments
