+++ monitor/api/src/main/java/org/datacleaner/monitor/referencedata/model/DictionaryModel.java
@@ -17,7 +17,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.datacleaner.monitor.server.controllers.referencedata.model;
+package org.datacleaner.monitor.referencedata.model;
 
 import java.util.List;
